requirements.txt
Flask==2.3.2
Flask==2.2.3
Flask_Bcrypt==1.0.1
mysql_connector_repackaged==0.3.1
PyJWT==2.7.0
PyJWT==1.7.1
PyJWT==2.8.0
python-dotenv==1.0.0
python_bcrypt==0.3.2
Requests==2.31.0
